san_interceptors: strncmp: call ENSURE_ASAN_INITED()
asan's overloaded strncmp does not ensure that ASAN is
initialized.  This causes a segfault when asan's strncmp is
used.

Fixed.

This patch is a subset of the patch commmited by upstream in
llvm r160853. Please see
http://code.google.com/p/address-sanitizer/issues/detail?id=94
for additional details.

Since r160853 still has a bug, I'd like to commit this patch as
is, knowing that it won't cause problems when we resync with
upstream.

Change-Id: Ic7d79d9ce90748fb01eb5190bc02a23e96e36cc3