Update patch set 1
Patch Set 1:

(1 comment)

Modulo the change to ubsan_init.cc, this actually worked out of the box. The output in logcat is a little strange, because it makes multiple calls to Printf to print the diagnostic, each of which calls into liblog, causing each piece to print on a separate line.

Not sure if we want to rewrite the calls to build the string and then print once, or rewrite AndroidLogPrint to buffer until it reaches a newline since we might run in to the same problem with the other sanitizers. The former sounds cleaner to me.

Patch-set: 1
Reviewer: Gerrit User 1043845 <1043845@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0