Remove invalid VFP variants of __eqdf2vfp/__eqsf2vfp.
https://code.google.com/p/android-developer-preview/issues/detail?id=2530

The VFP variants of these routines return 1 in the case of equality
(instead of 0, like the non-VFP cases). This causes issues for devices
that rely on software emulation of the floating point comparisons. This
went unnoticed since all prior devices we tested/shipped had hardware
units (and thus didn't need these routines). Upstream compiler-rt
doesn't use these routines by default, so it is sufficient to just
remove our remapping of __eqdf2 -> __eqdf2vfp (and similarly for the
single precision variant).

Change-Id: I626febf8b29f76128ed27f21c544179ce2d6c349