Support building libasan.a and libasan_cxx.a properly.
We previously only built shared objects for the device, so this allows
us to build (and then ship) static libraries for linking with device
code. This patch also cleans up the naming of libasan-preinit (which was
previously erroneously marked libasan), and also removes some now
useless LOCAL_ADDITIONAL_DEPENDENCY lines.

Change-Id: I662c260b3d4142b2ef098906efb0af6899dd4ace