Update patch set 2
Patch Set 2:

> > I'm not sure static linking of asan runtime library on device
 > would
 > > work correctly because of the linker differencies from linux. It
 > > probably would though.
 > > Also, static target libasan would inflate the system image size
 > w/o
 > > any real gains, so it would probably stay unused.
 > 
 > It shouldn't be part of the system image as a static library. Is
 > the intention always that the user has a matched dynamic library on
 > any given device? How can that work for the NDK, where we don't
 > even know if a device shipped any ASan-enabled runtime libraries? I
 > guess I am trying to clean this up NDK-wise.

Oh, the NDK absolutely requires a shared library.

If you just want to build a standalone executable with ASan, static libasan should work just fine.

The interesting case is a JNI application. Assuming the device does not include ASan (which is and probably will be the common case in the future) to run such an application we would need to ensure that ASan is initialized at the start of the process. This is done by LD_PRELOAD-ing the asan runtime library into the Zygote. NDK includes asan_device_setup.sh script for that. This does not work with the static libasan, of course.

I guess there is no harm in including the static libasan in the NDK, but the clang driver does not even know what to do with it. It always picks the shared runtime on android.

Patch-set: 2