DO NOT SUBMIT: Initialize Android logging.
Without this ubsan failures only go to stderr, not logcat.

The message is written to logcat all funky (broken up into several
lines). IIRC when I looked in to that about a year ago it was because
the write of the diagnostic is done as several writes and logd does
not buffer and wait for a newline, but prints immediately.

It might be better to drop the liblog dependency entirely and just
move the logging over to syslog (which will end up in logcat anyway).
We would need to dlopen liblog and use that for pre-L (M?).

Change-Id: I70d259030ff07733ec5dc6d61ace71f4cf606c6f