Update patch set 4
Patch Set 4:

Colin is right that this still leaves a bunch of *_chk() functions as dependencies inside of this libcompiler_rt.so. We need to work on reducing that down so that the resulting binary is indeed able to run on API 9 devices, but it is no worse than the present day builds that also contain those symbols (being that this is not using the NDK at all right now).

Patch-set: 4