asanwrapper: remove all environment editing.
LD_LIBRARY_PATH that asanwrapper used to set is out-of-date.

The use case of "asanwrapper app_process" does not need LD_LIBRARY_PATH at all,
because asan/app_process is built with asan and the linker magic happens.

The use case of asanwrapper with a non-asan binary is not supported (and never was),
because it requires LD_PRELOAD.

ASAN_OPTIONS is unnecessary because the same setting can be found in /system/asan.options.

Bug: 36202343

Change-Id: Ic3ac252c2cf84efdeb6ca39dee4480ae68e01374