Disable LTO to work around gold bug.
When using LTO, libcompiler_rt-extras.a gets built as LLVM bitcode. Some
modules that need this archive only refer to its symbols from LLVM bitcode, so
the linker doesn't know that it needs to pass the archive to the plugin until it
is too late. We are simply disabling LTO for the couple of source files in
libcompiler_rt-extras.a.

Test: Full build with LTO enabled
Change-Id: I87149a15333e48ef854ef9dc1866b9df3180a0e1