Update patch set 1
Patch Set 1: Code-Review+1

I think this is ok, but we really are working to remove this use of compiler-rt throughout Android (as we switch to a proper prebuilt for everything). I'll let Pirama and/or Yi comment on whether they have hit this with LTO yet too.

Patch-set: 1
Reviewer: Gerrit User 1010056 <1010056@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1