Update patch set 1
Patch Set 1:

We are now using the llvm-toolchain branch (and CMake build rules) to build Clang for Android.  To checkout this, you can 'repo init' with '-b llvm-toolchain'.

FWIW, we just generated prebuilts based on r305632.  But the next update will most likely be newer than 313360, so I don't think there's a need to cherry-pick these to llvm-toolchain branch.

 > Test: python external/clang/build.py
 > 
 > These changes should also be cherry-picked to master branch.
 > @Stephen Hines, do you want me to upload them there as well? I
 > believe, only the revert change should be omitted when
 > cherry-picking to master.

No need to cherry-pick to the master branch as they are not used to generate sanitizer runtimes (even for the target).

Patch-set: 1
CC: Gerrit User 1058860 <1058860@85c56323-6fa9-3386-8a01-6480fb634889>