Update patch set 1
Patch Set 1:

Are you planning to build/ship this library with the NDK? As it is, I am not sure how we could let the NDK depend on this. I also would have preferred to just start linking/shipping a working libclang_rt.builtins as part of the NDK build. This code (in external/) is really outdated, and unlikely to get updated in the short-term, so I am trying to understand how this is going to get used (i.e. we were trying to reduce dependencies here, but this is adding a new one).

Patch-set: 1
Reviewer: Gerrit User 1010056 <1010056@85c56323-6fa9-3386-8a01-6480fb634889>