Update patch set 1
Patch Set 1:

We don't have compiler-rt builtins for Windows yet (created b/128692407).  So this CL is needed as a stopgap.

As for compiler-rt, we seem to have switched only the device modules over and still use libgcc for Linux host modules.  Since we do package compiler-rt builtins for Linux, we should use that instead of the compiler-rt from this project.

Patch-set: 1
Reviewer: Gerrit User 1058860 <1058860@85c56323-6fa9-3386-8a01-6480fb634889>